+++ R/mltools.R
@@ -179,6 +179,11 @@ autostrat <- function(x,
 #'     the format of returned object.
 #' @param unique [default TRUE] to return only unique combinations of c1 and c2.
 #'     Set to FALSE to return both c1 > c2 and c2 > c1.
+#' @param func [default list()] (for advanced use only) a named list of
+#'     functions which take 2 arguments: 'core' and 'xlen', the coredata matrix
+#'     of the ichimoku object and the number of observations, respectively. Each
+#'     function must return a vector of length 'xlen', and these are included in
+#'     the grid.
 #'
 #' @return A data.frame or matrix in a 'tidy' format with one observation per
 #'     row and one feature per column with the target 'y' as the first column
@@ -225,7 +230,8 @@ mlgrid <- function(x,
                    dir = c("long", "short"),
                    type = c("boolean", "numeric", "z-score"),
                    format = c("dataframe", "matrix"),
-                   unique = TRUE) {
+                   unique = TRUE,
+                   func = list()) {
 
   is.ichimoku(x) || stop("mlgrid() only works on ichimoku objects", call. = FALSE)
   y <- match.arg(y)
@@ -252,6 +258,12 @@ mlgrid <- function(x,
 
   pairs <- .mlgrid_pairs
   veclist <- writeVectors(x = core, pairs = pairs, p2 = p2, xlen = xlen, type = type)
+
+  if (length(func)) {
+    for (i in seq_along(func))
+      func[[i]] <- func[[i]](core, xlen)
+    veclist <- c(func, veclist)
+  }
 
   if (!missing(unique) && !isTRUE(unique)) {
     pairs <- list(pairs[[2L]], pairs[[1L]])