+++ tests/testthat/test-mltools.R
@@ -1,7 +1,8 @@
 cloud <- ichimoku(sample_ohlc_data)
 grid <- mlgrid(cloud, y = "none", type = "numeric")
 grid2 <- mlgrid(cloud, y = "logret", type = "boolean", dir = "short", unique = FALSE)
-grid3 <- mlgrid(cloud, y = "ret", k = 3, type = "z-score", format = "matrix")
+grid3 <- mlgrid(cloud, y = "ret", k = 3, type = "z-score", format = "matrix",
+                func = list(close = function(core, xlen) core[1:xlen, "close"]))
 stratlist <- autostrat(cloud, n = 2, dir = "short", quietly = TRUE)
 
 test_that("autostrat ok", {
@@ -17,7 +18,9 @@ test_that("mlgrid ok", {
   expect_s3_class(grid2, "data.frame")
   expect_identical(dim(grid2), c(153L, 75L))
   expect_type(grid3, "double")
-  expect_identical(dim(grid3), c(151L, 38L))
+  expect_identical(dim(grid3), c(151L, 39L))
+  expect_length(expect_type(look(grid3)$means, "double"), 38L)
+  expect_length(expect_type(look(grid3)$sdevs, "double"), 38L)
   expect_warning(mlgrid(cloud, k = NA), regexp = "'k' invalid")
   expect_error(mlgrid(sample_ohlc_data), regexp = "ichimoku object")
 })